Manage delete confirmation box

+++ client/src/pages/DepartmentList.js
@@ -6,10 +6,10 @@ import { useNavigate } from 'react-router-dom';
 import { InputText } from 'primereact/inputtext';
 import { Dropdown } from 'primereact/dropdown';
 import { Button } from 'primereact/button';
-import { CButton } from '@coreui/react';
 import { useAuth } from '../context/AuthContext';
 import Layout from './Layout';
 import Loader from '../components/Loader';
+import { ConfirmDialog, confirmDialog } from 'primereact/confirmdialog';
 
 const DepartmentList = ({ title }) => {
   const { getDepartment, deleteDepartment } = useDepartment();
@@ -22,7 +22,6 @@ const DepartmentList = ({ title }) => {
   const [globalFilterValue, setGlobalFilterValue] = useState('');
   const [sortField, setSortField] = useState('createdAt');
   const [sortOrder, setSortOrder] = useState(-1);
-  const [visible, setVisible] = useState(false);
   const navigate = useNavigate();
 
   const fetchDepartments = async (currentPage, rowsPerPage, query, sortField, sortOrder) => {
@@ -51,46 +50,17 @@ const DepartmentList = ({ title }) => {
     fetchDepartments(1, rowsPerPage, globalFilterValue, sortField, sortOrder)
   };
 
-  const clear = () => {
-    deleteTost.current.clear();
-    setVisible(false);
-  };
-
-  const departmentDelete = async (id) => {
-    await deleteDepartment(id);
-    fetchDepartments(currentPage, rowsPerPage, globalFilterValue, sortField, sortOrder);
-    clear()
-  }
-
-  const handleDelete = async (id) => {
-
-    if (!visible) {
-      setVisible(true);
-      deleteTost.current.clear();
-      deleteTost.current.show({
-        severity: 'info',
-        sticky: true,
-        content: (
-          <div className="flex flex-column align-items-center" style={{ flex: '1'}}>
-            <div className="text-center">
-              <i className="pi pi-exclamation-triangle" style={{ fontSize: '3rem' }}></i>
-              <div className="font-bold text-xl my-3">Are you sure you want to delete this department?</div>
-            </div>
-            <div class="text-end">
-              <CButton color="info" onClick={(e) => clear(false)} className='ms-3'>No</CButton>
-              <CButton color="success" className="ms-3" onClick={() => departmentDelete(id)}>Yes</CButton>
-            </div>
-          </div>
-        )
-      });
-    }
-    // const confirmDelete = window.confirm(
-    //   'Are you sure you want to delete this department?'
-    // );
-    // if (confirmDelete) {
-    //   await deleteDepartment(id);
-    //   fetchDepartments(currentPage, rowsPerPage, globalFilterValue, sortField, sortOrder);
-    // }
+  const handleDelete = (id) => {
+    confirmDialog({
+      message: 'Are you sure you want to delete this department?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      position: 'top',
+      accept: async () => {
+        await deleteDepartment(id);
+        fetchDepartments(currentPage, rowsPerPage, globalFilterValue, sortField, sortOrder);
+      },
+    });
   };
 
   const handleUpdate = async (id) => {
@@ -119,6 +89,7 @@ const DepartmentList = ({ title }) => {
         <Loader />
       ) : (
         <>
+          <ConfirmDialog />
           <div className="card mb-5">
             <div className="mainHeader d-flex align-items-center justify-content-between">
               <div>
